Checking for subtitles when kodi isn't playing
Hi,

This fix add the ability for checking subtitles even when kodi isn't
playing.
There are two options for using it:
* 1 - By Manual Search
* 2 - Standing on an item (In Libraries or others known video addons)

In order to open the Subtitles Search dialog, you should need add it as
a key shortcut.
I also created context menu addon in order to do that.

I tested it and everything is working well, and there is no regression
from the last version.
This fix doesn't touch the regular operation as it is today.

- This fix already used in other subtitles addons (Hebrew subtitles
sites)

Thanks.